+++ tests/AppBundle/Controller/BlogControllerTest.php
@@ -13,6 +13,9 @@ namespace Tests\AppBundle\Controller;
 
 use AppBundle\Entity\Post;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+use Symfony\Component\HttpFoundation\Response;
+use Tests\ControllerTestTrait;
+use Tests\FixturesTrait;
 
 /**
  * Functional test for the controllers defined inside BlogController.
@@ -26,9 +29,12 @@ use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
  */
 class BlogControllerTest extends WebTestCase
 {
+    use ControllerTestTrait;
+    use FixturesTrait;
+
     public function testIndex()
     {
-        $client = static::createClient();
+        $client = $this->getAnonymousClient();
         $crawler = $client->request('GET', '/en/blog/');
 
         $this->assertCount(
@@ -40,7 +46,7 @@ class BlogControllerTest extends WebTestCase
 
     public function testRss()
     {
-        $client = static::createClient();
+        $client = $this->getAnonymousClient();
         $crawler = $client->request('GET', '/en/blog/rss.xml');
 
         $this->assertSame(
@@ -53,5 +59,36 @@ class BlogControllerTest extends WebTestCase
             $crawler->filter('item'),
             'The xml file displays the right number of posts.'
         );
+    }
+
+    /**
+     * This test changes the database contents by creating a new comment. However,
+     * thanks to the DAMADoctrineTestBundle and its PHPUnit listener, all changes
+     * to the database are rolled back when this test completes. This means that
+     * all the application tests begin with the same database contents.
+     */
+    public function testNewComment()
+    {
+        $client = $this->getUserClient();
+
+        /** @var Post $post */
+        $post = $client->getContainer()->get('doctrine')->getRepository(Post::class)->find(1);
+        $commentContent = $this->getRandomCommentContent();
+        $commentsCount = $post->getComments()->count();
+
+        $crawler = $client->request('GET', '/en/blog/posts/'.$post->getSlug());
+        $form = $crawler->selectButton('Publish comment')->form([
+            'comment[content]' => $commentContent,
+        ]);
+        $client->submit($form);
+
+        $this->assertSame(Response::HTTP_FOUND, $client->getResponse()->getStatusCode());
+
+        $post = $client->getContainer()->get('doctrine')->getRepository(Post::class)->find(1);
+        // The first one is the last inserted (See descending order of comments association).
+        $comment = $post->getComments()->first();
+
+        $this->assertSame($commentsCount + 1, $post->getComments()->count());
+        $this->assertSame($commentContent, $comment->getContent());
     }
 }
